// coach-portal/lib/types.ts
export type InviteStatus = "pending" | "accepted" | "revoked";
// ...
export type GetTeamInviteInfoResponse = {
  teamName: string;
  coachName: string;
  playerName: string;
  playerEmail: string;
  status: InviteStatus;
};